+++ packages/taro-transformer-wx/src/render.ts
@@ -28,8 +28,8 @@ import {
   parseJSXElement
 } from './jsx'
 import { DEFAULT_Component_SET, MAP_CALL_ITERATOR, LOOP_STATE, LOOP_CALLEE, THIRD_PARTY_COMPONENTS, LOOP_ORIGINAL, INTERNAL_GET_ORIGNAL } from './constant'
-import generate from 'babel-generator'
 import { Adapter } from './adapter'
+import generate from 'babel-generator'
 const template = require('babel-template')
 
 type ClassMethodsMap = Map<string, NodePath<t.ClassMethod | t.ClassProperty>>
@@ -220,7 +220,7 @@ export class RenderParser {
         const block2 = buildBlockElement()
         block.children = [consequent]
         newJSXIfAttr(block, test)
-        setJSXAttr(block2, 'wx:else')
+        setJSXAttr(block2, Adapter.else)
         block2.children = [alternate]
         const parentBlock = buildBlockElement()
         parentBlock.children = [block, block2]
@@ -241,7 +241,7 @@ export class RenderParser {
         block.children = [t.jSXExpressionContainer(consequent)]
         newJSXIfAttr(block, test)
         const block2 = buildBlockElement()
-        setJSXAttr(block2, 'wx:else')
+        setJSXAttr(block2, Adapter.else)
         block2.children = [t.jSXExpressionContainer(alternate)]
         const parentBlock = buildBlockElement()
         parentBlock.children = [block, block2]
@@ -260,7 +260,10 @@ export class RenderParser {
     const properties: t.ObjectProperty[] = []
     this.componentProperies.forEach((propName) => {
       properties.push(
-        t.objectProperty(t.stringLiteral(propName), t.nullLiteral())
+        t.objectProperty(t.stringLiteral(propName), t.objectExpression([
+          t.objectProperty(t.stringLiteral('type'), t.nullLiteral()),
+          t.objectProperty(t.stringLiteral('value'), t.nullLiteral())
+        ]))
       )
     })
     let classProp = t.classProperty(
@@ -361,7 +364,7 @@ export class RenderParser {
                       this.referencedIdentifiers.add(ary)
                     }
                   }
-                  setJSXAttr(jsxElementPath.node, 'wx:for', t.jSXExpressionContainer(ary))
+                  setJSXAttr(jsxElementPath.node, Adapter.for, t.jSXExpressionContainer(ary))
                   this.loopCalleeId.add(findFirstIdentifierFromMemberExpression(callee))
                   const [func] = callExpr.node.arguments
                   if (
@@ -372,7 +375,7 @@ export class RenderParser {
                     if (t.isIdentifier(item)) {
                       setJSXAttr(
                         jsxElementPath.node,
-                        'wx:for-item',
+                        Adapter.forItem,
                         t.stringLiteral(item.name)
                       )
                       this.loopScopes.add(item.name)
@@ -381,14 +384,14 @@ export class RenderParser {
                     } else {
                       setJSXAttr(
                         jsxElementPath.node,
-                        'wx:for-item',
+                        Adapter.forItem,
                         t.stringLiteral('__item')
                       )
                     }
                     if (t.isIdentifier(index)) {
                       setJSXAttr(
                         jsxElementPath.node,
-                        'wx:for-index',
+                        Adapter.forIndex,
                         t.stringLiteral(index.name)
                       )
                       this.loopScopes.add(index.name)
@@ -475,7 +478,7 @@ export class RenderParser {
                 if (parentIfStatement) {
                   setJSXAttr(
                     jsxElementPath.node,
-                    'wx:elif',
+                    Adapter.elseif,
                     t.jSXExpressionContainer(test),
                     jsxElementPath
                   )
@@ -483,7 +486,7 @@ export class RenderParser {
                   if (this.topLevelIfStatement.size > 0) {
                     setJSXAttr(
                       jsxElementPath.node,
-                      'wx:elif',
+                      Adapter.elseif,
                       t.jSXExpressionContainer(test),
                       jsxElementPath
                     )
@@ -494,7 +497,7 @@ export class RenderParser {
                 }
               }
             } else if (block.children.length !== 0) {
-              setJSXAttr(jsxElementPath.node, 'wx:else')
+              setJSXAttr(jsxElementPath.node, Adapter.else)
             }
             block.children.push(jsxElementPath.node)
             this.finalReturnElement = block
@@ -515,7 +518,7 @@ export class RenderParser {
               if (isBlockIfStatement(ifStatement, blockStatement)) {
                 const { test, alternate, consequent } = ifStatement.node
                 if (alternate === blockStatement.node) {
-                  setJSXAttr(jsxElementPath.node, 'wx:else')
+                  setJSXAttr(jsxElementPath.node, Adapter.else)
                 } else if (consequent === blockStatement.node) {
                   const parentIfStatement = ifStatement.findParent(p =>
                     p.isIfStatement()
@@ -523,7 +526,7 @@ export class RenderParser {
                   if (parentIfStatement && parentIfStatement.get('alternate') === ifStatement) {
                     setJSXAttr(
                       jsxElementPath.node,
-                      'wx:elif',
+                      Adapter.elseif,
                       t.jSXExpressionContainer(test),
                       jsxElementPath
                     )
@@ -633,18 +636,18 @@ export class RenderParser {
               const element = p.get('openingElement') as NodePath<t.JSXOpeningElement>
               if (element.get('name').isJSXIdentifier({ name: 'block' })) {
                 const attrs = element.node.attributes
-                const hasWXForLoop = attrs.some(attr => t.isJSXIdentifier(attr.name, { name: 'wx:for' }))
-                const hasWXKey = attrs.some(attr => t.isJSXIdentifier(attr.name, { name: 'wx:key' }))
+                const hasWXForLoop = attrs.some(attr => t.isJSXIdentifier(attr.name, { name: Adapter.for }))
+                const hasWXKey = attrs.some(attr => t.isJSXIdentifier(attr.name, { name: Adapter.key }))
                 return hasWXForLoop && !hasWXKey
               }
             }
             return false
           }) as NodePath<t.JSXElement>
           if (loopBlock) {
-            setJSXAttr(loopBlock.node, 'wx:key', value)
+            setJSXAttr(loopBlock.node, Adapter.key, value)
             path.remove()
           } else {
-            path.get('name').replaceWith(t.jSXIdentifier('wx:key'))
+            path.get('name').replaceWith(t.jSXIdentifier(Adapter.key))
           }
         } else if (
           name.name.startsWith('on')
@@ -1044,10 +1047,10 @@ export class RenderParser {
                 }
               }
             })
-            // setJSXAttr(returned, 'wx:for', t.identifier(stateName))
+            // setJSXAttr(returned, Adapter.for, t.identifier(stateName))
             this.addRefIdentifier(callee, t.identifier(stateName))
             // this.referencedIdentifiers.add(t.identifier(stateName))
-            setJSXAttr(component.node, 'wx:for', t.jSXExpressionContainer(t.identifier(stateName)))
+            setJSXAttr(component.node, Adapter.for, t.jSXExpressionContainer(t.identifier(stateName)))
             const returnBody = this.renderPath.node.body.body
             const ifStem = callee.findParent(p => p.isIfStatement())
             // @TEST